done with one note
-everything works fine on simulator,
but on a device (iPhone 4a with iOS9.3) it is not possible to remove tv
event from favorites or watchlist - in both cases same request is sent
to server but different resposes are received_

+++ MovieApp/MovieApp/TVEvent.h
@@ -18,7 +18,7 @@
 @property (nonatomic) BOOL isInFavorites;
 @property (nonatomic) BOOL isInWatchlist;
 @property (nonatomic) BOOL isInRatings;
-@property (nonatomic) float rating;
+@property (nonatomic) NSInteger rating;
 @property (nonatomic) NSInteger duration;
 
 -(NSString *)getGenreNameForId:(NSUInteger)genreId;
